refactor: channel 도메인에서 controller, dto, service 클래스에서 버전별로 분리
