+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/request/v1/SendSignalRequestDto.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.domain.channel.dto.request;
+package com.hertz.hertz_be.domain.channel.dto.request.v1;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
