+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/request/v2/SignalMatchingRequestDto.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.domain.channel.dto.request;
+package com.hertz.hertz_be.domain.channel.dto.request.v2;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
