+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/response/v1/ChannelListResponseDto.java
@@ -1,4 +1,4 @@
-package com.hertz.hertz_be.domain.channel.dto.response;
+package com.hertz.hertz_be.domain.channel.dto.response.v1;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
