+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/dto/response/v1/SendSignalResponseDto.java
@@ -1,3 +1,3 @@
-package com.hertz.hertz_be.domain.channel.dto.response;
+package com.hertz.hertz_be.domain.channel.dto.response.v1;
 
 public record SendSignalResponseDto (Long channelRoomId) {}
