+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/service/v1/ChannelService.java
@@ -1,8 +1,9 @@
-package com.hertz.hertz_be.domain.channel.service;
+package com.hertz.hertz_be.domain.channel.service.v1;
 
 import com.corundumstudio.socketio.SocketIOServer;
-import com.hertz.hertz_be.domain.channel.dto.request.SignalMatchingRequestDto;
-import com.hertz.hertz_be.domain.channel.dto.response.*;
+import com.hertz.hertz_be.domain.channel.dto.response.v1.*;
+import com.hertz.hertz_be.domain.channel.service.AsyncChannelService;
+import com.hertz.hertz_be.domain.channel.service.SseChannelService;
 import com.hertz.hertz_be.domain.user.responsecode.UserResponseCode;
 import com.hertz.hertz_be.global.common.NewResponseCode;
 import com.hertz.hertz_be.global.exception.*;
@@ -10,7 +11,7 @@ import com.hertz.hertz_be.global.socketio.dto.SocketIoMessageResponse;
 import com.hertz.hertz_be.domain.channel.entity.*;
 import com.hertz.hertz_be.domain.channel.entity.enums.Category;
 import com.hertz.hertz_be.domain.channel.entity.enums.MatchingStatus;
-import com.hertz.hertz_be.domain.channel.dto.request.SendSignalRequestDto;
+import com.hertz.hertz_be.domain.channel.dto.request.v1.SendSignalRequestDto;
 import com.hertz.hertz_be.domain.channel.responsecode.*;
 import com.hertz.hertz_be.domain.channel.repository.*;
 import com.hertz.hertz_be.domain.channel.repository.projection.ChannelRoomProjection;
@@ -45,7 +46,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Slf4j
-@Service
+@Service("channelServiceV1")
 @RequiredArgsConstructor
 public class ChannelService {
 
@@ -58,7 +59,6 @@ public class ChannelService {
     private final UserInterestsRepository userInterestsRepository;
     private final SignalRoomRepository signalRoomRepository;
     private final SignalMessageRepository signalMessageRepository;
-    private final ChannelRoomRepository channelRoomRepository;
     private final InterestsService interestsService;
     private final AsyncChannelService asyncChannelService;
     private final WebClient webClient;
@@ -84,7 +84,6 @@ public class ChannelService {
         this.userInterestsRepository = userInterestsRepository;
         this.signalMessageRepository = signalMessageRepository;
         this.signalRoomRepository = signalRoomRepository;
-        this.channelRoomRepository = channelRoomRepository;
         this.interestsService = interestsService;
         this.asyncChannelService = asyncChannelService;
         this.aesUtil = aesUtil;
@@ -560,68 +559,5 @@ public class ChannelService {
                 asyncChannelService.sendNewMessageNotifyToPartner(signalMessage, partnerId, false);
             }
         });
-    }
-
-    @Transactional
-    public String channelMatchingStatusUpdate(Long userId, SignalMatchingRequestDto response, MatchingStatus matchingStatus) {
-        SignalRoom room = signalRoomRepository.findById(response.getChannelRoomId())
-                .orElseThrow(() -> new BusinessException(
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getCode(),
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getHttpStatus(),
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getMessage()
-                ));
-
-        int updatedSender = signalRoomRepository.updateSenderMatchingStatus(room.getId(), userId, matchingStatus);
-        int updatedReceiver = signalRoomRepository.updateReceiverMatchingStatus(room.getId(), userId, matchingStatus);
-
-        if (updatedSender == 0 && updatedReceiver == 0) {
-            throw new BusinessException(
-                    ChannelResponseCode.ALREADY_EXITED_CHANNEL_ROOM.getCode(),
-                    ChannelResponseCode.ALREADY_EXITED_CHANNEL_ROOM.getHttpStatus(),
-                    "사용자가 참여 중인 채팅방 아닙니다."
-            );
-        }
-
-        entityManager.flush();
-
-
-        TransactionSynchronizationManager.registerSynchronization(new TransactionSynchronization() {
-            @Override
-            public void afterCommit() {
-                asyncChannelService.notifyMatchingResultToPartner(room, userId, matchingStatus);
-                asyncChannelService.createMatchingAlarm(room, userId);
-            }
-        });
-
-
-        /**
-         * 매칭 수락/거절 후 현재 관계 반환
-         */
-        if(matchingStatus == MatchingStatus.MATCHED) {
-            return signalRoomRepository.findMatchResultByUser(userId, room.getId());
-        } else {
-            return ChannelResponseCode.MATCH_REJECTION_SUCCESS.getCode();
-        }
-    }
-
-
-    @Transactional
-    public void leaveChannelRoom(Long roomId, Long userId) {
-        SignalRoom room = signalRoomRepository.findById(roomId)
-                .orElseThrow(() -> new BusinessException(
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getCode(),
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getHttpStatus(),
-                        ChannelResponseCode.CHANNEL_NOT_FOUND.getMessage()
-                ));
-
-        if (!userRepository.existsById(userId)) {
-            throw new BusinessException(
-                    UserResponseCode.USER_NOT_FOUND.getCode(),
-                    UserResponseCode.USER_NOT_FOUND.getHttpStatus(),
-                    "채팅방 나가기 요청한 사용자가 존재하지 않습니다."
-            );
-        }
-
-        room.leaveChannelRoom(userId);
     }
 }
