+++ hertz-be/src/main/java/com/hertz/hertz_be/domain/channel/service/v3/ChannelService.java
@@ -1,0 +1,11 @@
+package com.hertz.hertz_be.domain.channel.service.v3;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Slf4j
+@Service("channelServiceV3")
+@RequiredArgsConstructor
+public class ChannelService {
+}
